requirements.txt
Django==4.2
gunicorn
tensorflow==2.17.1
numpy
pillow
whitenoise
